Texture2D FluidSim2D
Created a FluidSim2D test for testing Jos Stam fluid dynamics. Wasn't too sure how to display the fluid so created Texture2D test to test how to create OpenGL texture and manipulate the pixels in the texture. Right now the texture2D test is displaying a grid of cells that can be manipulated to display different coloured cells. I'll try and use this approach to display the density field in the FluidSim2D test.