Updated fluid simulation test environment
Added missing diffuse for y axis. Now fluid has better fluid dynamics and actually looks like a fluid. Added addFluid controls for the ImGui interface. Can now inject density into the middle of the fluid field with the ImGui gui and also change which way the velocity of the fluid travels in with ImGui interface.